+++ src/components/subs/TeamBuilder.tsx
@@ -210,7 +210,10 @@ const TeamBuilder: React.FC<{
               skillLevel: 1,
               masterRank: 0,
               level: maxLevel[card.rarity],
-              trained: card.rarity >= 3,
+              trained:
+                card.cardRarityType !== "rarity_birthday" && card.rarity >= 3,
+              trainable:
+                card.cardRarityType !== "rarity_birthday" && card.rarity >= 3,
               story1Unlock: true,
               story2Unlock: true,
             },
@@ -441,9 +444,29 @@ const TeamBuilder: React.FC<{
             const maxNormalLevel = [0, 20, 30, 50, 60];
             return Object.assign({}, state, {
               masterRank: 0,
-              trained: state.level > maxNormalLevel[card.rarity],
+              trained:
+                card.cardRarityType !== "rarity_birthday" &&
+                state.level > maxNormalLevel[card.rarity],
+              trainable:
+                card.cardRarityType !== "rarity_birthday" && card.rarity >= 3,
               story1Unlock: false,
               story2Unlock: false,
+            });
+          }
+        );
+      } else if (
+        !Object.prototype.hasOwnProperty.call(
+          currentEntry.teamCardsStates[0],
+          "trainable"
+        )
+      ) {
+        currentEntry.teamCardsStates = currentEntry.teamCardsStates.map(
+          (state) => {
+            const card = cards!.find((elem) => elem.id === state.cardId)!;
+
+            return Object.assign({}, state, {
+              trainable:
+                card.cardRarityType !== "rarity_birthday" && card.rarity >= 3,
             });
           }
         );
@@ -506,7 +529,7 @@ const TeamBuilder: React.FC<{
   );
 
   const handleLoadSekaiTeamEntry = useCallback(() => {
-    if (!sekaiProfile || !sekaiProfile.sekaiUserProfile) return;
+    if (!sekaiProfile || !sekaiProfile.sekaiUserProfile || !cards) return;
 
     const cardIds = Array.from({ length: 5 }).map(
       (_, idx) =>
@@ -520,11 +543,16 @@ const TeamBuilder: React.FC<{
         (state) => state.cardId === cardId
       )!;
 
+      const cardInfo = cards.find((card) => card.id === cardId)!;
+      const trainable =
+        cardInfo.cardRarityType !== "rarity_birthday" && cardInfo.rarity >= 3;
+
       return {
         cardId,
         level: userCard.level,
         masterRank: userCard.masterRank,
-        trained: userCard.specialTrainingStatus === "done",
+        trained: trainable && userCard.specialTrainingStatus === "done",
+        trainable,
         skillLevel: 1,
         story1Unlock: userCard.episodes[0].scenarioStatus !== "unreleased",
         story2Unlock: userCard.episodes[1].scenarioStatus !== "unreleased",
@@ -539,6 +567,7 @@ const TeamBuilder: React.FC<{
     setLoadTeamDialogVisible(false);
   }, [
     calcTotalPower,
+    cards,
     sekaiProfile,
     setTeamCards,
     setTeamCardsStates,
@@ -958,6 +987,10 @@ const TeamBuilder: React.FC<{
                                   `member${idx + 1}` as "member1"
                                 ]
                               }
+                              trained={
+                                sekaiProfile.sekaiUserProfile!.userCards[idx]
+                                  .specialTrainingStatus === "done"
+                              }
                             />
                           </Grid>
                         ))}
@@ -1022,7 +1055,10 @@ const TeamBuilder: React.FC<{
                             xs={4}
                             sm={2}
                           >
-                            <CardThumb cardId={cardId} />
+                            <CardThumb
+                              cardId={cardId}
+                              trained={team.teamCardsStates[0].trained}
+                            />
                           </Grid>
                         ))}
                       </Grid>
@@ -1120,13 +1156,15 @@ const TeamBuilder: React.FC<{
                   fullWidth
                 />
               </Grid>
-              <Grid item>
-                <FormControlLabel
-                  control={<Switch checked={editingCard.trained} />}
-                  label={t("card:trained")}
-                  onChange={(e, checked) => handleChange(checked, "trained")}
-                />
-              </Grid>
+              {editingCard.trainable && (
+                <Grid item>
+                  <FormControlLabel
+                    control={<Switch checked={editingCard.trained} />}
+                    label={t("card:trained")}
+                    onChange={(e, checked) => handleChange(checked, "trained")}
+                  />
+                </Grid>
+              )}
               <Grid item>
                 <FormControlLabel
                   control={<Switch checked={editingCard.story1Unlock} />}