+++ CHANGELOG.md
@@ -2,6 +2,44 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.8.0](https://github.com/Sekai-World/sekai-viewer/compare/v0.7.2...v0.8.0) (2021-10-18)
+
+
+### Features
+
+* add separate settings page ([e0b7f72](https://github.com/Sekai-World/sekai-viewer/commits/e0b7f727ed6c8cc9411e6dd74b26dcec26f4912c))
+* **event-tracker:** support multiple server region ([59d45c9](https://github.com/Sekai-World/sekai-viewer/commits/59d45c946b1d2e938e6d9d40657e3b7a23fd9d31))
+* support multiple server region ([64c307d](https://github.com/Sekai-World/sekai-viewer/commits/64c307dcc6e6e52707c23816cda4c48bbe51d49d))
+* **utils:** support multiple server regions ([38ceec3](https://github.com/Sekai-World/sekai-viewer/commits/38ceec3cd121f8cf2dee6fdd29d6c003b10bf9ca))
+
+
+### Refactors
+
+* **support:** remove container ([86cf630](https://github.com/Sekai-World/sekai-viewer/commits/86cf6303d2f4b6eece09b4b443495e5bee8f9783))
+
+### [0.7.2](https://github.com/Sekai-World/sekai-viewer/compare/v0.7.1...v0.7.2) (2021-10-04)
+
+
+### Bug Fixes
+
+* **card:** card birthday frame ([9064146](https://github.com/Sekai-World/sekai-viewer/commits/9064146aefae5a342fba8138b5efc2a9d45e81c6))
+* **card:** card detail birthday card thumb ([b581e6a](https://github.com/Sekai-World/sekai-viewer/commits/b581e6a8eeb548a5ce902c1ef63ab42a9e1d5637))
+* **subs:** birthday card, trainable state ([e88a2e9](https://github.com/Sekai-World/sekai-viewer/commits/e88a2e981831042077f3801c3956e6e53b01f74a))
+* **subs:** correctly handle thumb training status ([f3bbcf6](https://github.com/Sekai-World/sekai-viewer/commits/f3bbcf648c51d231a1ba27dec1633b894ddf45bb))
+* **subs:** loadTeamDialog show card training status ([cbd69d8](https://github.com/Sekai-World/sekai-viewer/commits/cbd69d8c5227249faa7b8dfac5946f7181dfe289))
+
+### [0.7.1](https://github.com/Sekai-World/sekai-viewer/compare/v0.7.0...v0.7.1) (2021-10-04)
+
+
+### Features
+
+* **card:** support birthday card ([f190891](https://github.com/Sekai-World/sekai-viewer/commits/f190891e40beceaa5cdf758df37e0141633f2af2))
+
+
+### Bug Fixes
+
+* **home:** haruka birthday banner date check ([e10db33](https://github.com/Sekai-World/sekai-viewer/commits/e10db33fe44dae6bf09f09a541d709a72c96b170))
+
 ## [0.7.0](https://github.com/Sekai-World/sekai-viewer/compare/v0.6.0...v0.7.0) (2021-10-03)